added current date by redefining sql db, optimized scrapes, and some formatting
